[UM-20]-created-student-report

+++ src/main/java/com/archisacadeny/student/CourseStudentMapper.java
@@ -18,7 +18,7 @@ public class CourseStudentMapper {
                             id INTEGER DEFAULT nextval('s_c_mapper_id') PRIMARY KEY NOT NULL,
                             student_id INTEGER NOT NULL,
                             course_id INTEGER NOT NULL,
-                            grade FLOAT ,
+                            grade DOUBLE PRECISION,
                             course_start_date DATE,
                             course_end_date DATE,
                             attended_lessons INTEGER,
@@ -38,6 +38,7 @@ public class CourseStudentMapper {
 
     public void saveToCourseStudentMapper(int studentID, int courseId, double grade, Timestamp start, Timestamp end, int attendedLessons, int missedLessons ){
         String query = "INSERT INTO course_student_mapper(student_id,course_id,grade,course_start_date,course_end_date,attended_lessons,missed_lessons) VALUES(?,?,?,?,?,?,?)";
+
         try(PreparedStatement statement = DataBaseConnectorConfig.getConnection().prepareStatement(query)){
             statement.setInt(1,studentID);
             statement.setInt(2,courseId);
@@ -45,8 +46,7 @@ public class CourseStudentMapper {
             statement.setTimestamp(4,start);
             statement.setTimestamp(5,end);
             statement.setInt(6,attendedLessons);
-
-
+          
             statement.execute();
 
         }catch(SQLException e){
